유효성 검사 일부 추가

+++ src/main/java/com/sparta/springcore/controller/PostController.java
@@ -2,11 +2,15 @@ package com.sparta.springcore.controller;
 
 import com.sparta.springcore.dto.PostRequestDto;
 import com.sparta.springcore.model.Post;
+import com.sparta.springcore.model.ResponseResult;
 import com.sparta.springcore.security.UserDetailsImpl;
 import com.sparta.springcore.service.PostService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.ui.Model;
+import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
@@ -45,14 +49,22 @@ public class PostController {
 
     // 게시글 등록
     @PostMapping("/api/posts")
-    public void createPost(@RequestBody PostRequestDto requestDto) {
+    public ResponseEntity<ResponseResult> createPost(@RequestBody PostRequestDto requestDto, Errors errors) {
+        if (errors.hasErrors()) {
+            return new ResponseEntity<>(new ResponseResult("fail", null), HttpStatus.OK);
+        }
         postService.createPost(requestDto);
+        return new ResponseEntity<>(new ResponseResult("success", null), HttpStatus.OK);
     }
 
     // 게시글 삭제
     @DeleteMapping("/api/posts/{postNo}")
-    public void deletePost(@PathVariable Integer postNo) {
+    public ResponseEntity<ResponseResult> deletePost(@PathVariable Integer postNo, Errors errors) {
+        if (errors.hasErrors()) {
+            return new ResponseEntity<>(new ResponseResult("fail", null), HttpStatus.OK);
+        }
         postService.deletePost(postNo);
+        return new ResponseEntity<>(new ResponseResult("success", null), HttpStatus.OK);
     }
 
     // 게시글 수정
